requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.11.0
pydantic-settings>=2.2.0
pydantic-extra-types>=2.0.0
pydantic_br>=1.1.0
email-validator>=1.1.3
PyJWT>=2.7.0
sentry-sdk>=2.29.1
structlog>=25.0.0
bcrypt>=4.0.1
jinja2>=3.1.0
requests>=2.32.3
cryptography>=43.0.0
python_multipart>=0.0.9
httpx>=0.27
pytest>=8
pytest-asyncio>=0.23
ruff>=0.5
redis>=5.0.0
pyserial
